requirements.txt
langchain-google-genai
python-dotenv
pandas
tabulate
langchain-community
langchain_experimental
matplotlib